gatekeeper/pyproject.toml
[project]
name = "gatekeeper"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "arekkulczycki", email = "ten.arek@gmail.com"},
]
dependencies = [
    "flask-pydantic>=0.12.0",
    "pytest-sqlalchemy-mock>=0.1.5",
    "gunicorn>=21.2.0",
    "pydantic-settings>=2.2.1",
    "psycopg2-binary>=2.9.9",
    "Flask-SQLAlchemy>=3.1.1",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
